requirements.txt
altair==4.2.2
attrs==22.2.0
blinker==1.6.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
entrypoints==0.4
filelock==3.9.0
fonttools==4.39.3
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
importlib-metadata==6.3.0
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
mpmath==1.2.1
networkx==3.0
numpy==1.24.1
packaging==23.1
pandas==1.5.3
Pillow==9.3.0
protobuf==3.20.3
pyarrow==11.0.0
pydeck==0.8.0
Pygments==2.15.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
rich==13.3.4
six==1.16.0
smmap==5.0.0
streamlit==1.21.0
sympy==1.11.1
toml==0.10.2
toolz==0.12.0
torch==2.0.0
torchaudio==2.0.1
torchutils==0.0.4
torchvision==0.15.1
tornado==6.2
typing_extensions==4.4.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.13
validators==0.20.0
watchdog==3.0.0
zipp==3.15.0
